+++ library/src/schemas/tuple/tuple.ts
@@ -1,18 +1,22 @@
 import type { BaseSchema, ErrorMessage, Issues, Pipe } from '../../types.ts';
-import { executePipe, getIssues, getSchemaIssues } from '../../utils/index.ts';
+import {
+  executePipe,
+  getIssues,
+  getRestAndDefaultArgs,
+  getSchemaIssues,
+} from '../../utils/index.ts';
 import type { TupleOutput, TupleInput, TuplePathItem } from './types.ts';
-import { getTupleArgs } from './utils/index.ts';
 
 /**
  * Tuple shape type.
  */
-export type TupleShape = [BaseSchema, ...BaseSchema[]];
+export type TupleItems = [BaseSchema, ...BaseSchema[]];
 
 /**
  * Tuple schema type.
  */
 export type TupleSchema<
-  TTupleItems extends TupleShape,
+  TTupleItems extends TupleItems,
   TTupleRest extends BaseSchema | undefined = undefined,
   TOutput = TupleOutput<TTupleItems, TTupleRest>
 > = BaseSchema<TupleInput<TTupleItems, TTupleRest>, TOutput> & {
@@ -28,7 +32,7 @@ export type TupleSchema<
  *
  * @returns A tuple schema.
  */
-export function tuple<TTupleItems extends TupleShape>(
+export function tuple<TTupleItems extends TupleItems>(
   items: TTupleItems,
   pipe?: Pipe<TupleOutput<TTupleItems, undefined>>
 ): TupleSchema<TTupleItems>;
@@ -42,7 +46,7 @@ export function tuple<TTupleItems extends TupleShape>(
  *
  * @returns A tuple schema.
  */
-export function tuple<TTupleItems extends TupleShape>(
+export function tuple<TTupleItems extends TupleItems>(
   items: TTupleItems,
   error?: ErrorMessage,
   pipe?: Pipe<TupleOutput<TTupleItems, undefined>>
@@ -58,8 +62,8 @@ export function tuple<TTupleItems extends TupleShape>(
  * @returns A tuple schema.
  */
 export function tuple<
-  TTupleItems extends TupleShape,
-  TTupleRest extends BaseSchema
+  TTupleItems extends TupleItems,
+  TTupleRest extends BaseSchema | undefined
 >(
   items: TTupleItems,
   rest: TTupleRest,
@@ -77,8 +81,8 @@ export function tuple<
  * @returns A tuple schema.
  */
 export function tuple<
-  TTupleItems extends TupleShape,
-  TTupleRest extends BaseSchema
+  TTupleItems extends TupleItems,
+  TTupleRest extends BaseSchema | undefined
 >(
   items: TTupleItems,
   rest: TTupleRest,
@@ -87,8 +91,8 @@ export function tuple<
 ): TupleSchema<TTupleItems, TTupleRest>;
 
 export function tuple<
-  TTupleItems extends TupleShape,
-  TTupleRest extends BaseSchema
+  TTupleItems extends TupleItems,
+  TTupleRest extends BaseSchema | undefined = undefined
 >(
   items: TTupleItems,
   arg2?: Pipe<TupleOutput<TTupleItems, TTupleRest>> | ErrorMessage | TTupleRest,
@@ -96,7 +100,7 @@ export function tuple<
   arg4?: Pipe<TupleOutput<TTupleItems, TTupleRest>>
 ): TupleSchema<TTupleItems, TTupleRest> {
   // Get rest, error and pipe argument
-  const [rest, error, pipe] = getTupleArgs<
+  const [rest, error, pipe] = getRestAndDefaultArgs<
     TTupleRest,
     Pipe<TupleOutput<TTupleItems, TTupleRest>>
   >(arg2, arg3, arg4);
@@ -128,11 +132,7 @@ export function tuple<
      */
     _parse(input, info) {
       // Check type of input
-      if (
-        !Array.isArray(input) ||
-        (!rest && items.length !== input.length) ||
-        (rest && items.length > input.length)
-      ) {
+      if (!Array.isArray(input) || items.length > input.length) {
         return getSchemaIssues(
           info,
           'type',
@@ -186,7 +186,7 @@ export function tuple<
       }
 
       // If necessary parse schema of each rest item
-      if (rest) {
+      if (rest && !(info?.abortEarly && issues)) {
         for (let key = items.length; key < input.length; key++) {
           const value = input[key];
           const result = rest._parse(value, info);