+++ library/src/schemas/tuple/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './getTupleArgs/index.ts';